Extract editor classes out of test namespace and add anchor support
